remove element tables, quote ebml element definitions, intro new mkv element tables
This is a complex commit, but it removes the existing element tables to
prepare for a subsequent commit with adds a process to a makefile to
generate and add tables created directly from the
matroska_ebml_schema.xml file. I also removed the matroska definitions
and instead quote and then supplement the same definitions from ebml.