make more elements deprecated via maxver attributes
This commit brings this specification in line with the old one on
www.matroska.org. The elements shown as dark gray text on light gray
background where all meant to be deprecated. Unfortunately the XML file
in the matroska-foundation repository did not reflect this fact.

As I couldn't find out when exactly each of those elements had been
deprecated I chose to use version 3 as the maximum version.