Adjust definition of "forced" and "default" flags to better reflect usage
This avoids implying user-hostile behavior that doesn't match actual
implementations (that I'm aware of), and is more in-line with the common
usage of "forced" subtitles for translations of parts of an audio or video
track that are in different languages than their primary language tag.

It also avoids requiring an overlay of multiple subtitle tracks. While this
was a good idea in theory (to avoid duplication of events between a "forced"
track and a normal one), it doesn't seem to have been implemented in players
or authoring software.