assume values not found in WebM have their default value
WebM doesn't use all our elements but sometimes they are implied even though
they don't define it. For example ContentEncodingOrder.

The elements must not be in WebM files (for compliance) but a parser may need
the actual value (of ContentEncodingOrder) to make any sense. So we assume
elements have their default values if they are not found.

If this is too loose, we may have to list the elements not in WebM but which
default value should be used.