Targets shouldn't be mandatory
There is currently a mismatch in the definition of `Targets`: On the one hand, it is mandatory (minOccurs="1"), while on the other hand the description describes what happens when this element is not present. This conflict is resolved by keeping the description and changing the syntax, because this saves space for the not uncommon use-case in which the tag applies to the whole segment. Furthermore, changing it in this way doesn't invalidate any existing files, whereas changing the description to match the syntax would potentially invalidate files that abide by the current description and not the syntax.