XML-escape HTML elements in <documentation> nodes
The documentation uses HTML tags such as <a> for links, <br/> for line
breaks or <strong> for emphasis. Unfortunately those were not
XML-escaped. That meant that XML parsers determined <a href=…> to be a
whole new child node of the <description> instead of being part of the
text of the <description>.

In order to reproduce the HTML-equivalent of such child ones,
transformations or programs would have to iterate recursively over all
child nodes of <documentation>, iterate over all existing attributes
etc. Very tedious and superfluous.