Add support for zstd content compression
The other compression schemes supported in Matroska suffer from having to be
applied on a per-packet level, with no global state. This means that they can't
exploit any of the redundancy between packets, and any huffman tables or other
configuration has to be duplicated in every packet. This severely limits the
attainable compression efficiency to approximately the level attained by zlib,
which has led to the other algorithms receiving very little adoption by content
authors, and little support from tool vendors.

The zstd library, on the other hand, supports generating a dictionary from a
large number of small inputs (e.g. packets), and using that dictionary to
compress similar inputs more efficiently. This improves performance on subtitle
inputs dramatically.

Muxing tools can either scan the input and pass its packets into zstd before
muxing an individual file, or provide an auxiliary tool that generates a
dictionary from the packets of one or more input files, then saves the result
to a file that can be reused when muxing tracks with similar content.