mark recurring elements with a maxOccurs or 1
In the past there was some confusion between elements that could be repeated and elements that could be multiple.

The EBML specs makes it clear that recurring elements are found multiple times (of the same copy) and that maxOccurs is ignored in that case.

adding the maxOccurs value makes it clearer that there cannot be multiple variants of this element, only exact copies.