Remove default value of newly added flags
These new flags all have a default value of zero and are not mandatory.
This means that according to the new EBML rules, the default value will
only matter when encountering an Empty Element. Yet according to the old
rules, setting the default value means that the element is always
present (if its parent is present) even if absent which means that all
tracks in old Matroska files would then be interpreted as e.g. not being
in their original language.

This is a potential source for confusion, in particular given that
knowledge of the new rules does not seem widespread. And it can be
easily fixed by removing the default values: According to the new rules,
Empty integer Elements with a default value of zero will be interpreted
the same as Empty Elements without a default value, so this changes
nothing for parsers adhering to the new specification.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>