mark strings with no default value as not usable if empty
Wether it's a mandatory or non-mandatory element, if a string has no default
value, and it's written in the file it should have a value, otherwise don't write
it.

For mandatory an empty string could be a way cancel the default value. But we
allow it for elements with a default value.

It may be legal in EBML to write empty strings for some reason, but
we don't have the use case in Matroska.