don't output the EBML elements from the Schema
They are already explained in the "Added Constraints on EBML" section.

The elements definitions are now part of the generic "Matroska Schema" section.
This makes the list of top level elements easier to read in the (HTML) summary.

A level of elements was removed because xml2rfc cannot handle more than 6 levels.