make ChapterDisplay use only one language at a time
Following the discussion in https://github.com/ietf-wg-cellar/ebml-specification/issues/409
It could be assumed that English is always present because ChapLanguage is
mandatory and has a default value.

To avoid this interpretation we should have one language (and one IETF variant)
per ChapterDisplay item. We already have to use many ones per language. The merge
of different values because they have the same value is unlikely (but it may
exist in the real world).