Remove range attribute from element with a (large) restriction section
The restriction mentions the value that are allowed. The range gives the lower
and upper limits. But without an indication on what to do with values in the
range but with an unknown value we have more confusion than necessary.

If the restriction values are later expanded older parser will ignore the new
values no matter what. The range is only useful to tell if the value is legit
or not, but there's still plenty of changes it is not.