make AlphaMode mandatory
This is a WebM thing loosely defined.

Having just the presence of the element present to signal something is at odd
with all other elements. The default value in this case has no use. And
following the discussions in #501 elements with a default value would preferably
be mandatory. But that's not possible with the "presence" signaling something.

Both libavformat [1] and libwebm [2] write a value of 1 when they want to signal
it.

Chromium requires a value of 1 to consider the track has alpha [3].
Firefox consider there's no alpha value is 0 is stored [4].

[1] https://github.com/FFmpeg/FFmpeg/blob/master/libavformat/matroskaenc.c#L1274
[2] https://aomedia.googlesource.com/aom/+/refs/heads/master/third_party/libwebm/mkvmuxer/mkvmuxer.cc#1522
[3] https://github.com/chromium/chromium/blob/72ceeed2ebcd505b8d8205ed7354e862b871995e/media/formats/webm/webm_video_client.cc#L149
[4] https://github.com/mozilla/nestegg/blob/ec6adfbbf979678e3058cc4695257366f39e290b/src/nestegg.c#L2627