simplify BlockAddIDType value 0
If the BlockAddIDType is 0, its default value, the value of the `BlockAddIDType`
is actually the `BlockAddIDValue` value, which is the `BlockAddID` value found in `BlockAdditions`.

It can only be 1 as otherwise the type would come from the BlockAdditionMapping, which is then 0 (the BlockAddIDType value)

It corresponds to the legacy mode that was originally intended with this value.