cropping: make `DisplayWidth`/`DisplayHeight` include cropped area
The primary motivation for this change is to ensure that Matroska files
with `PixelCrop*` elements are displayed consistently, regardless of
whether cropping is applied. This makes such files compatible across
software that either supports or ignores these elements. Without this
change, files suffer from limited compatibility, whereas cropping tags
should enhance the experience rather than be a requirement for proper
image display.

A secondary motivation is that an overwhelming number of files in the
wild do not actually adjust the `DisplayWidth`/`DisplayHeight` elements
as expected by the specification. Most of the time, these values are
left uncropped, with the expectation that cropping will be applied
later. This issue arises partly because MKVToolNix does not
automatically perform this calculation, it only sets the `PixelCrop*`
elements unless explicitly instructed by the user. Observing the
existing files in circulation, it is evident that almost no one manually
adjusts these values.

Lastly, some subtitle formats depend on the video size for positioning
and rendering, making them sensitive to cropping. Currently, most
authoring tools do not take cropping into account. By making the
Matroska specification more flexible and allowing subtitles to be
cropped along with the video, it becomes easier to produce files that
are compatible with various Matroska players.

For a more detailed discussion on this topic, please refer to the links
below:

See: https://gitlab.com/mbunkus/mkvtoolnix/-/issues/2389
See: https://github.com/mpv-player/mpv/pull/13446